Viewport버그, UI 입력 버그를 수정하고  text관리자를 생성해라
UI버그를 수정하는 도중, Viewport에도 버그가 있다는 것을 깨달았다.
Viewport의 화면을 늘리게 되면 화면이 잘려서 나오는등의 자잘한 버그가 있었는데
camera에서 texture에 그릴 때 Viewport 및 ScissorRect설정을 하지 않아서 일어난
버그였다.
물론 윈도우 Resize를 현재로서는 지원을 하지 않을 계획이라 최종 Texture만
Resize를 하게 되서 UI및 카메라 Texture들은 그만큼 비율이 확대, 축소되서 그려진다.

UI의 버튼 입력이 약간 맞지 않는 버그가 있다. 윈도우 자체의 TitleBar에 있는
그 영역때문에 밀려나서 그런 것이다 . 그래서 코드를 확인해 봤더니 윈도우 전체 크기로
Render Target및 Camera의 Texture등을 생성하고 있었다. 그렇기 때문에 전체적으로 Y축에서 오차가
심한 모습을 보였던 것이다. 그래서 그 값들을 뺀 값으로 ID3D12Resource들을 생성하게 설정하였다.

UI의 Hovered Event실행 관련으로 제일 위에 Hovered된 UI만 동작하게 해야하는데
지금은 그 밑에 있는 UI까지 Hovered Event가 발생한다. 따라서 ui event를 수정하였는데
먼저 기존 상속 관계가 너무 제약적이라고 생각해서 상속 관계를 지우고 (UI Event Info)
hovered event와 unhovered event를 생성하고 hovered event가 실행이 된 경우라면 그 뒤에 나오는
hovered event를 모두 무시, 그 외의 상태에선 unhovered event가 실행하게 설정하였다.

기존의 text render system에서 text render manager를 파생시켰는데, brush나 text format은
따로 관리하는 것이 좋을 것 같아서 이렇게 분리하였다.

+++ Client/ClientFW/src/client/renderer/renderitem/ui_render_item.cpp
@@ -52,7 +52,7 @@ namespace client_fw
 			}
 		}
 
-		m_num_of_ui_layer = ui_layers.size();
+		m_num_of_ui_layer = static_cast<UINT>(ui_layers.size());
 
 		if (m_num_of_ui_layer > 0)
 		{
